pyproject.toml
[tool.poetry]
name = "hospital-project-django"
version = "0.1.0"
description = "The same project as Hospital_Project but with django"
authors = ["Kiki Tsepe <kiki.tsepe@gmail.com>"]
readme = "README.md"
packages = [{include = "hospital_project_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
django-filter = "^24.1"
django-filter-stubs = "^0.1.3"
fabric = "^3.2.2"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
pydantic = "^2.6.4"
drf-spectacular = "^0.27.1"
stubs = "^1.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
yapf = "^0.40.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
